+++ handpose/network.py
@@ -128,3 +128,130 @@ class TransferNetwork(torch.nn.Module):
         
         """
         torch.save(self.model.state_dict(), network_path)
+
+def network_head(pred, 
+                 require_kpt_conf,
+                 S,
+                 B,
+                 nkpt,
+                 nc
+                ):
+    """Returns the head of the network.
+
+    For bounding box ``x y w h``, each of these tensor is of size ``(m, B, S, S)``.
+    All the boxes values are stored in their respective type of annotation.
+    
+    Parameters
+    ----------
+    pred: torch.Tensor 
+        A tensor of size ``(m, N)`` 
+        where ``m`` is the number of samples
+        and ``N`` is the number of features.
+
+        .. math::
+            N = S \\times S \\times (B \\times (5 + nkpt_{dim} \\times nkpt) + nc) \\\\
+    require_kpt_conf: bool
+        Use keypoint confidence.
+    S: int
+        The grid size.
+    B: int
+        The number of prediction boxes.
+    nkpt: int
+        The number of landmarks.
+    nc: int
+        Number of classes to predict
+        
+    Returns
+    -------
+    dict
+        A dictionary with keys ``['conf', 'x', 'y', 'w', 'h', 'k_conf', 'kpt', 'classes']``
+    
+    Examples
+    --------
+    >>> pred = torch.randn((2, 7 * 7 * (2 * (5 + 3 * 21) + 2)))
+    >>> head = network_head(pred=pred, require_kpt_conf=True, S=7, B=2, nkpt=21, nc=2)
+    
+    """
+
+    # Selecting the keypoint tensor dimension based on whether keypoint confidence are considered
+    nkpt_dim = 2
+
+    if require_kpt_conf:
+        nkpt_dim = 3
+
+    try:
+        m, _ = pred.shape
+    except Exception as e:
+        print(e)
+        raise
+
+    tensor_ch = B * (5 + nkpt_dim * nkpt) + nc
+    
+    # Reshaping
+    tensor = pred.reshape((m, tensor_ch, S, S))
+    
+    start = 0
+    end = start+B
+    conf = tensor[:,start:end, ...]
+
+    start = end
+    end = start+B
+    x = tensor[:, start:end, ...]
+
+    start = end
+    end = start + B
+    y = tensor[:, start:end, ...]
+
+    start = end
+    end = start + B
+    w = tensor[:, start:end, ...]
+
+    start = end
+    end = start + B
+    h = tensor[:, start:end, ...]
+    
+    start = end
+    end = start + nc
+    class_scores = tensor[:, start:end,... ]
+
+    start = end
+    kpt_tensor = tensor[:, start:, ...]
+
+    kpt_conf_dict = dict()
+    kpt_dict = dict()
+
+    i = 0
+    j = 0
+    while i < (nkpt * 2):
+        start = i
+        end = i + B
+        kx = kpt_tensor[:,start:end, ...]
+        ky = kpt_tensor[:,end:end+B, ...]
+
+        kpt_dict[f'kx_{j}'] = kx
+        kpt_dict[f'ky_{j}'] = ky
+
+        i += 2
+        j += 1
+        
+
+    if require_kpt_conf:
+        kpt_conf_tensor = kpt_tensor[:,-nkpt*B:,...]
+        k = 0
+        l = 0
+        while k < (nkpt * B):
+            kpt_conf_dict[f'k_conf_{l}'] = kpt_conf_tensor[:,k:k+B,...]
+            k += B
+            l += 1
+
+    head = {'conf': conf,
+            'x': x,
+            'y': y,
+            'w': w,
+            'h': h,
+            'k_conf': kpt_conf_dict,
+            'kpt': kpt_dict,
+            'classes': class_scores
+           }
+    
+    return head